requirements.txt
APScheduler==3.6.3
blinker==1.4
boto3==1.13.1
botocore==1.16.1
certifi==2019.3.9
chardet==3.0.4
click==7.1.1
cmdstanpy==0.4.0
convertdate==2.2.0
cycler==0.10.0
Cython==0.29.16
dateparser==0.7.4
docutils==0.15.2
ephem==3.7.7.1
fbprophet==0.6
Flask==1.1.2
holidays==0.10.2
idna==2.9
itsdangerous==1.1.0
jenkins==1.0.2
jenkinsapi==0.3.11
Jinja2==2.11.2
jmespath==0.9.5
kiwisolver==1.2.0
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
lunardate==0.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
multi-key-dict==2.0.3
numpy==1.18.3
pandas==1.0.3
pbr==5.5.0
prometheus-api-client==0.2.0
prometheus-client==0.7.1
PyMeeus==0.3.7
pyparsing==2.4.7
pystan==2.19.1.1
python-dateutil==2.8.1
python-jenkins==1.7.0
pytz==2019.3
PyYAML==5.3.1
redis==3.4.1
regex==2020.4.4
requests==2.23.0
retrying==1.3.3
s3transfer==0.3.3
schedule==0.6.0
sentry-sdk==0.10.2
setuptools-git==1.2
six==1.14.0
tzlocal==2.0.0
urllib3==1.25.9
Werkzeug==1.0.1
